requirements.txt
requests>=2.31.0
apprise==1.9.1
websocket-client>=1.6.0